Backend user setup complete

+++ backend/src/services/recipe_posts.js
@@ -1,6 +1,6 @@
 import { Recipe } from '../db/models/recipe_posts.js'
-export async function createRecipe({title, description, ingredients, steps, author, tags}){
-    const newRecipe = new Recipe({ title, description, ingredients, steps, author, tags })
+export async function createRecipe(userId, {title, description, ingredients, steps, tags}){
+    const newRecipe = new Recipe({ title, description, ingredients, steps, author: userId, tags })
     return await newRecipe.save()
 } 
     
@@ -22,13 +22,13 @@ export async function listRecipesByTag(tags, options) {
 export async function getRecipeById(recipeId) {
   return await Recipe.findById(recipeId)
 }
-export async function updateRecipe(recipeId, { title, description, ingredients, steps, author, tags }) {
+export async function updateRecipe(userId, recipeId, { title, description, ingredients, steps, tags }) {
   return await Recipe.findOneAndUpdate(
-    { _id: recipeId },
-    { $set: { title, description, ingredients, steps, author, tags } },
+    { _id: recipeId, author:userId },
+    { $set: { title, description, ingredients, steps, tags } },
     { new: true },
   )
 }
-export async function deleteRecipe(recipeId) {
-  return await Recipe.deleteOne({ _id: recipeId })
+export async function deleteRecipe(recipeId,userId) {
+  return await Recipe.deleteOne({ _id: recipeId, author: userId  })
 }
